+++ src/main/java/main/controller/ItemStockContoller.java
@@ -1,10 +1,12 @@
 package main.controller;
 
+import main.model.auth.AuthServiceSession;
 import main.model.db.dao.ItemStockDAO;
 import main.model.db.dto.db.ItemStockDTO;
 import main.model.db.dto.itemStockList.ItemStockRequest;
 import main.model.db.dto.itemStockList.JoinedItemStockDTO;
 
+import main.properties.CustomProperties;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,10 +22,20 @@ public class ItemStockContoller {
     @Autowired
     private ItemStockDAO itemStockDAO;
 
+    @Autowired
+    private AuthServiceSession authServiceSession;
+
+    @Autowired
+    private CustomProperties customProperties;
+
     @PostMapping("/list")
     public List<JoinedItemStockDTO> getItemStockList(
-            @RequestBody ItemStockRequest request,
+            @RequestBody(required = false) ItemStockRequest request,
             @RequestParam(required = false) String state) {
+
+        if(!authServiceSession.getSessionUser().equals(request.getAffiliationCode()) || !authServiceSession.getSessionUser().equals(customProperties.getAffiliationCode())) {
+            return null;
+        }
 
         if (state == null || state.isEmpty()) {
             return itemStockDAO.getItemStockList(request.getAffiliationCode());
@@ -34,9 +46,10 @@ public class ItemStockContoller {
 
     @PostMapping("/listAll")
     public List<JoinedItemStockDTO> getAllItemStockList(
-            @RequestBody ItemStockRequest request,
             @RequestParam(required = false) String state) {
-
+        if(!authServiceSession.getSessionUser().equals(customProperties.getAffiliationCode())) {
+            return null;
+        }
         if (state == null || state.isEmpty()) {
             return itemStockDAO.getAllItemStockList();
         } else {
@@ -48,6 +61,9 @@ public class ItemStockContoller {
     //ToDo : 약간의 검토 필요(물자를 추가할때 입고일이 다르다고 분립을 시킬것인지 기존 데이터에 병합시킬것인지 만약 병합시키지 않는다면 표시하는쪽에서 나눠서 표시할것인지 합쳐서 표기할것인지)
     @PostMapping("/add")
     public String addItemStock(@RequestBody ItemStockDTO itemStock) {
+        if(!authServiceSession.getSessionUser().equals(itemStock.getAffiliationCode()) || !authServiceSession.getSessionUser().equals(customProperties.getAffiliationCode())) {
+            return "권한이 없습니다.";
+        }
         try {
             int result = itemStockDAO.insertItemStock(itemStock);
             return result > 0 ? "물자추가 성공" : "추가 실패";
@@ -62,6 +78,9 @@ public class ItemStockContoller {
     public String updateItemStock(@RequestParam int stockId,
                                   @RequestParam(required = false) Integer quantity,
                                   @RequestParam(required = false) String status) {
+        if(!authServiceSession.getSessionUser().equals(itemStockDAO.findById(stockId).getAffiliationCode()) || !authServiceSession.getSessionUser().equals(customProperties.getAffiliationCode())) {
+            return "권한이 없습니다.";
+        }
         try {
             int result = 0;
 
@@ -83,6 +102,9 @@ public class ItemStockContoller {
     // 삭제 (Delete)
     @GetMapping("/delete")
     public String deleteItemStock(@RequestParam int stockId) {
+        if(!authServiceSession.getSessionUser().equals(itemStockDAO.findById(stockId).getAffiliationCode()) || !authServiceSession.getSessionUser().equals(customProperties.getAffiliationCode())) {
+            return "권한이 없습니다.";
+        }
         try {
             int result = itemStockDAO.deleteItemStock(stockId);
             return result > 0 ? "삭제 성공" : "삭제 실패";
@@ -94,6 +116,9 @@ public class ItemStockContoller {
     // 재고 감소
     @GetMapping("/decrease")
     public String decreaseStock(@RequestParam int itemId, @RequestParam String affiliationCode, @RequestParam int quantity) {
+        if(!authServiceSession.getSessionUser().equals(affiliationCode) || !authServiceSession.getSessionUser().equals(customProperties.getAffiliationCode())) {
+            return "권한이 없습니다.";
+        }
         try {
             itemStockDAO.decreaseStock(itemId, affiliationCode, quantity);
             return "재고 감소 성공";
